feat(order): 결제 전 주문 정보 저장 기능 구현

+++ src/main/java/roomescape/reservationWaiting/service/ReservationWaitingService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.exception.AuthorizationException;
 import roomescape.order.domain.Order;
+import roomescape.order.exception.OrderAmountMismatchException;
+import roomescape.order.exception.OrderNotFoundException;
 import roomescape.order.repository.OrderRepository;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.exception.DuplicateReservationException;
@@ -115,7 +117,7 @@ public class ReservationWaitingService {
     }
 
     @Transactional
-    public void promoteWaiting(Long waitingId, String name) {
+    public void promoteWaiting(Long waitingId, String name, String orderId) {
         ReservationWaiting waiting = reservationWaitingRepository.findById(waitingId)
                 .orElseThrow(ReservationWaitingNotFoundException::new);
 
@@ -131,14 +133,19 @@ public class ReservationWaitingService {
             throw new ReservationWaitingNotPromotableException();
         }
 
+        Order order = orderRepository.findByOrderId(orderId)
+                .orElseThrow(OrderNotFoundException::new);
+
+        if (!order.getAmount().equals(waiting.getTheme().getAmount())) {
+            throw new OrderAmountMismatchException();
+        }
+
         int affectedRow = reservationWaitingRepository.deleteById(waiting.getId());
         int nonAffected = 0;
 
         if (affectedRow == nonAffected) {
             throw new ReservationWaitingNotFoundException();
         }
-
-        Order order = orderRepository.save(Order.of(waiting.getTheme().getAmount()));
 
         try {
             reservationRepository.save(
